[root] Access to the root
It turned out that the JS code for the hgui made
the assumption that the root / was served as actual
content. The web server unfortunately did a 
redirect to the actual file.

However, JS is nowadays quite brain damaged and
gets a tad lost when the root url becomes /index.html.

This patch allows a bundle to set Web-Root manifest
header to a file. If any bundle has this header,
it will be served for '/'

If there are multiple bundles with this header,
a warning is logged once and then one of them is 
chosen.